implemented an enemy script and player attack state
enemy script contains the enemy's current hp. it should be updated to contain fields for its damage and such as well.
player attack state gets references to the player and enemy game objects. then it uses them to retrieve the enemy's current hp, subtract the damage the player deals, and update the enemy's hp.
a player script should be implemented to contain information such as player hp, player damage, etc. rather than using a hard number in the player attack state script.
there are also debug log entries.

i also created player and enemy game objects in the scene (a plain capsule and circle, respectively). the enemy game object has the enemy script.

truthfully, i have been down a huge rabbit hole trying to understand this for over an hour. i think this is the simplest way to do it, but i might be wrong. unity forums are kind of rude and surprisingly beginner-hostile.